refactor(permanents): Major refactor of the submatrices calculators. Place the interface and the base classes into separate files. Prepare base class for Guan code-related submatrices calculators, that will take care of the boilerplate code.
