+++ theboss/boson_sampling_utilities/permanent_calculators/bs_cc_ch_submatrices_permanent_calculator.py
@@ -7,123 +7,31 @@ __author__ = "Tomasz Rybotycki"
     step of the algorithm. Instead of computing each permanent separately, we can
     compute them all in one run, which is vastly more efficient.
     
-    Instead of using Rysers formula, or Glynn formula, we base our approach on the
+    Instead of using Ryser's formula, or Glynn formula, we base our approach on the
     Chin-Huh's formula, which takes into account possible bunching in the input states
     (or rather can be interpreted as that). 
 """
 
-import abc
-from numpy import complex128, ndarray, int64, array, nonzero, zeros, ones
+from numpy import complex128, ndarray, array, nonzero, zeros_like
 import operator
 from functools import reduce
 from typing import List, Optional
 
-
-class BSSubmatricesPermanentCalculatorInterface(abc.ABC):
-    """
-        This is the interface class for submatrices permanents calculator. For now
-        BSCCCHSubmatricesPermanentCalculator is the only class of that kind, but it's
-        possible that there will be more (based e.g. on Glynns formula that grants
-        numerical stability.
-
-        If the above happens, place the interface class in the separate file.
-    """
-
-    @abc.abstractmethod
-    def compute_permanents(self) -> List[complex128]:
-        """Computes permanent of a matrix given before."""
-        raise NotImplementedError
-
-    @property
-    @abc.abstractmethod
-    def matrix(self) -> ndarray:
-        raise NotImplementedError
-
-    @matrix.setter
-    @abc.abstractmethod
-    def matrix(self, matrix: ndarray) -> None:
-        raise NotImplementedError
-
-    @property
-    @abc.abstractmethod
-    def input_state(self) -> ndarray:
-        raise NotImplementedError
-
-    @input_state.setter
-    @abc.abstractmethod
-    def input_state(self, input_state: ndarray) -> None:
-        raise NotImplementedError
-
-    @property
-    @abc.abstractmethod
-    def output_state(self) -> ndarray:
-        raise NotImplementedError
-
-    @output_state.setter
-    @abc.abstractmethod
-    def output_state(self, output_state: ndarray) -> None:
-        raise NotImplementedError
+from theboss.boson_sampling_utilities.permanent_calculators.bs_submatrices_permanent_calculator_base import (
+    BSGuanBasedSubmatricesPermanentCalculatorBase,
+)
 
 
-class BSSubmatricesPermanentCalculatorBase(
-    BSSubmatricesPermanentCalculatorInterface, abc.ABC
+class BSCCCHSubmatricesPermanentCalculator(
+    BSGuanBasedSubmatricesPermanentCalculatorBase
 ):
-    """
-        Base class for BSSubmatricesPermanentCalculator classes. It takes care of some
-        boilerplate code.
-
-        Again, it should be put into separate file were the
-        BSCCCHSubmatricesPermanentCalculator cease to be the only submatrices
-        permanent calculator.
-    """
-
-    def __init__(
-        self,
-        matrix: ndarray,
-        input_state: Optional[ndarray] = None,
-        output_state: Optional[ndarray] = None,
-    ) -> None:
-        if output_state is None:
-            output_state = array([], dtype=int64)
-        if input_state is None:
-            input_state = array([], dtype=int64)
-        self._matrix = matrix
-        self._input_state = input_state
-        self._output_state = output_state
-
-    @property
-    def matrix(self) -> ndarray:
-        return self._matrix
-
-    @matrix.setter
-    def matrix(self, matrix: ndarray) -> None:
-        self._matrix = matrix
-
-    @property
-    def input_state(self) -> ndarray:
-        return self._input_state
-
-    @input_state.setter
-    def input_state(self, input_state: ndarray) -> None:
-        self._input_state = array(input_state, dtype=int64)
-
-    @property
-    def output_state(self) -> ndarray:
-        return self._output_state
-
-    @output_state.setter
-    def output_state(self, output_state: ndarray) -> None:
-        self._output_state = array(output_state, dtype=int64)
-
-
-class BSCCCHSubmatricesPermanentCalculator(BSSubmatricesPermanentCalculatorBase):
 
     """
-        The name stands for Boson Sampling Clifford & Clifford Chin-Huh submatrices
-        permanent calculator, as it uses Clifford & Clifford approach to compute
-        permanents of submatrices to compute sub-distribution of Boson Sampling problem
-        instance. The starting point in our case is Chin-Huh permanent calculator
-        iterated in Guan Codes induced order.
+    The name stands for Boson Sampling Clifford & Clifford Chin-Huh submatrices
+    permanent calculator, as it uses Clifford & Clifford approach to compute
+    permanents of submatrices to compute sub-distribution of Boson Sampling problem
+    instance. The starting point in our case is Chin-Huh permanent calculator
+    iterated in Guan Codes induced order.
     """
 
     def __init__(
@@ -133,12 +41,10 @@ class BSCCCHSubmatricesPermanentCalculator(BSSubmatricesPermanentCalculatorBase)
         output_state: Optional[ndarray] = None,
     ) -> None:
 
-        self.sums: dict = dict()
-        self.permanents: List[complex128] = []
-        self.multiplier: int = 1
-        self.binomials_product: int = 1
-        self.v_vector: ndarray = array(0)
-        self.considered_columns_indices = array(0)
+        self._sums: dict = dict()
+        self.permanents: ndarray = array([], dtype=complex128)
+        self._multiplier: int = 1
+        self._considered_columns_indices = array(0)
 
         super().__init__(matrix, input_state, output_state)
 
@@ -147,116 +53,86 @@ class BSCCCHSubmatricesPermanentCalculator(BSSubmatricesPermanentCalculatorBase)
         # TODO TR:  This method is huge and complicated. It would be smart to break
         #           it down into smaller ones.
 
-        self.permanents = [complex128(0) for _ in range(len(self.input_state))]
+        self._initialize_permanents_computation()
 
-        # Required for Guan Code iteration
-        self.v_vector = zeros(len(self._input_state), dtype=int)  # g
-        code_update_information = ones(len(self._input_state), dtype=int)  # u
-        position_limits = list(self._input_state)  # n
+        while self._r_vector[-1] <= self._position_limits[-1]:
 
-        self.sums = dict()
-        self.binomials_product = 1
-        self.considered_columns_indices = nonzero(self._output_state)[0]
+            self._update_guan_code()
 
-        self.multiplier = 1
+            if self._index_to_update == len(self._r_vector):
+                self.permanents /= pow(2, sum(self.input_state) - 1)
+                return list(self.permanents)
 
-        # Initialization (0-th step).
-        for i in self.considered_columns_indices:
-            self.sums[i] = 0
-            for j in range(len(self._input_state)):
-                self.sums[i] += self._input_state[j] * self._matrix[i][j]
+            self._multiplier = -self._multiplier
 
-        self._add_permanent_addends()
-
-        # Rest of the steps.
-        while self.v_vector[-1] <= position_limits[-1]:
-
-            # UPDATE R VECTOR
-            index_to_update = 0  # i
-            updated_value_at_index = self.v_vector[0] + code_update_information[0]  # k
-            while (
-                updated_value_at_index > position_limits[index_to_update]
-                or updated_value_at_index < 0
-            ):
-                code_update_information[index_to_update] = -code_update_information[
-                    index_to_update
-                ]
-                index_to_update += 1
-
-                if index_to_update == len(self.v_vector):
-
-                    for _ in range(int(sum(self.input_state)) - 1):
-                        for i in range(len(self.permanents)):
-                            self.permanents[i] /= 2
-
-                    return self.permanents
-
-                updated_value_at_index = (
-                    self.v_vector[index_to_update]
-                    + code_update_information[index_to_update]
-                )
-
-            last_value_at_index = self.v_vector[index_to_update]
-            self.v_vector[index_to_update] = updated_value_at_index
-            # END UPDATE
-
-            # START PERMANENT UPDATE
-            self.multiplier = -self.multiplier
-
-            # Sums update
-            for i in self.sums:
-                self.sums[i] -= (
-                    2
-                    * (self.v_vector[index_to_update] - last_value_at_index)
-                    * self.matrix[i][index_to_update]
-                )
-
-            # Binoms update
-            if self.v_vector[index_to_update] > last_value_at_index:
-                self.binomials_product *= (
-                    self._input_state[index_to_update] - last_value_at_index
-                ) / self.v_vector[index_to_update]
-            else:
-                self.binomials_product *= last_value_at_index / (
-                    self._input_state[index_to_update] - self.v_vector[index_to_update]
-                )
-
+            self._update_sums()
+            self._update_binomials_product()
             self._add_permanent_addends()
 
         for _ in range(int(sum(self._input_state)) - 1):
-            for i in range(len(self.permanents)):
-                self.permanents[i] /= 2
+            self.permanents /= pow(2, sum(self.input_state) - 1)
 
-        return self.permanents
+        return list(self.permanents)
+
+    def _initialize_permanents_computation(self) -> None:
+        """
+        Initialize all the class fields prior to the permanents computation.
+        """
+        self.permanents = zeros_like(self.input_state, dtype=complex128)
+
+        self.initialize_guan_codes_variables()
+
+        self._sums = dict()
+        self._considered_columns_indices = nonzero(self._output_state)[0]
+
+        self._multiplier = 1
+
+        # Initialization (0-th step).
+        for i in self._considered_columns_indices:
+            self._sums[i] = 0
+            for j in range(len(self._input_state)):
+                self._sums[i] += self._input_state[j] * self._matrix[i][j]
+
+        self._add_permanent_addends()
+
+    def _update_sums(self) -> None:
+        """
+        Update sums instead of recomputing them.
+        """
+        for i in self._sums:
+            self._sums[i] -= (
+                2
+                * (self._r_vector[self._index_to_update] - self._last_value_at_index)
+                * self.matrix[i][self._index_to_update]
+            )
 
     def _add_permanent_addends(self) -> None:
+        """
+        Update the intermediate permanents.
+        """
         # For each occupied mode
         for i in range(len(self.input_state)):
 
-            if self.input_state[i] == 0 or self.input_state[i] == self.v_vector[i]:
+            if self.input_state[i] == 0 or self.input_state[i] == self._r_vector[i]:
                 continue
 
             # Compute update the sums
             updated_sums = {}
-            for j in self.considered_columns_indices:
-                updated_sums[j] = self.sums[j] - self.matrix[j][i]
+            for j in self._considered_columns_indices:
+                updated_sums[j] = self._sums[j] - self.matrix[j][i]
 
             # Compute update binomial product
-            updated_binom = self.binomials_product / (
-                self.input_state[i] / (self.input_state[i] - self.v_vector[i])
+            updated_binom = self._binomials_product / (
+                self.input_state[i] / (self.input_state[i] - self._r_vector[i])
             )
 
-            addend = (
-                self.multiplier
-                * updated_binom
-                * reduce(
-                    operator.mul,
-                    [
-                        pow(updated_sums[j], self._output_state[j])
-                        for j in self.considered_columns_indices
-                    ],
-                    1,
-                )
+            addend = reduce(
+                operator.mul,
+                [
+                    pow(updated_sums[j], self._output_state[j])
+                    for j in self._considered_columns_indices
+                ],
+                self._multiplier * updated_binom,
             )
 
             self.permanents[i] += addend
