+++ theboss/boson_sampling_utilities/permanent_calculators/bs_cc_ryser_submatrices_permanent_calculator.py
@@ -5,15 +5,18 @@ __author__ = "Tomasz Rybotycki"
     calculator 
 """
 
-from theboss.boson_sampling_utilities.permanent_calculators.bs_cc_ch_submatrices_permanent_calculator import (
+from theboss.boson_sampling_utilities.permanent_calculators.bs_submatrices_permanent_calculator_base import (
     BSSubmatricesPermanentCalculatorBase,
+    BSGuanBasedSubmatricesPermanentCalculatorBase,
 )
 
 from numpy import ndarray, complex128, nonzero, zeros, ones
-from typing import Optional, List, Union
+from typing import Optional, List
 
 
-class BSCCRyserSubmatricesPermanentCalculator(BSSubmatricesPermanentCalculatorBase):
+class BSCCRyserSubmatricesPermanentCalculator(
+    BSGuanBasedSubmatricesPermanentCalculatorBase
+):
 
     """
     This class computes the permanents of the submatrices using slightly modified
@@ -53,20 +56,9 @@ class BSCCRyserSubmatricesPermanentCalculator(BSSubmatricesPermanentCalculatorBa
         self._matrix_t = matrix.transpose()
 
         self._sums: dict = dict()  # w_j(r)
-        self._forward_products: Union[dict, None] = None
-        self._backward_products: Union[dict, None] = None
         self._considered_columns_indices = nonzero(self._output_state)[0]
         self._multiplier = 1
         self._binomials_product = 1
-
-        # Required for Guan Code iteration. Some values are also used during updates.
-        self._r_vector: ndarray = zeros(len(self._input_state), dtype=int)  # g
-        self._code_update_information: ndarray = ones(
-            len(self._input_state), dtype=int
-        )  # u
-        self._position_limits: List[int] = list(self._input_state)  # n
-        self._index_to_update: int = 0
-        self._last_value_at_index: int = 0
 
         self.permanents: List[complex128] = []
 
@@ -106,47 +98,14 @@ class BSCCRyserSubmatricesPermanentCalculator(BSSubmatricesPermanentCalculatorBa
         self.permanents = [complex128(0) for _ in range(len(self.input_state))]
 
         self._sums = dict()
-        self._forward_products = None
-        self._backward_products = None
         self._multiplier = pow(-1, sum(self.output_state))
         self._considered_columns_indices = nonzero(self._output_state)[0]
         self._binomials_product = 1
 
-        # Required for Guan Code iteration
-        self._r_vector = zeros(len(self._input_state), dtype=int)  # g
-        self._code_update_information = ones(len(self._input_state), dtype=int)  # u
-        self._position_limits = list(self._input_state)  # n
+        self.initialize_guan_codes_variables()
 
         for j in self._considered_columns_indices:
             self._sums[j] = 0
-
-    def _update_guan_code(self) -> None:
-        """
-        Prepare the class for processing the next Guan code.
-        """
-        self._index_to_update = 0  # i
-        updated_value_at_index = (
-            self._r_vector[0] + self._code_update_information[0]
-        )  # k
-        while (
-            updated_value_at_index > self._position_limits[self._index_to_update]
-            or updated_value_at_index < 0
-        ):
-            self._code_update_information[
-                self._index_to_update
-            ] = -self._code_update_information[self._index_to_update]
-            self._index_to_update += 1
-
-            if self._index_to_update == len(self._r_vector):
-                return
-
-            updated_value_at_index = (
-                self._r_vector[self._index_to_update]
-                + self._code_update_information[self._index_to_update]
-            )
-
-        self._last_value_at_index = self._r_vector[self._index_to_update]
-        self._r_vector[self._index_to_update] = updated_value_at_index
 
     def _update_permanents(self) -> None:
         """
@@ -188,18 +147,3 @@ class BSCCRyserSubmatricesPermanentCalculator(BSSubmatricesPermanentCalculatorBa
                 )
 
             self.permanents[i] += prod * updated_binomials_product
-
-    def _update_binomials_product(self) -> None:
-        """
-        Update the binomials product to reflect the new Guan code instead of recomputing
-        it.
-        """
-        if self._r_vector[self._index_to_update] > self._last_value_at_index:
-            self._binomials_product *= (
-                self._input_state[self._index_to_update] - self._last_value_at_index
-            ) / self._r_vector[self._index_to_update]
-        else:
-            self._binomials_product *= self._last_value_at_index / (
-                self._input_state[self._index_to_update]
-                - self._r_vector[self._index_to_update]
-            )
